Interview Stage: Practical Assessment – CRUD Application Development
The purpose of this assessment was to create a crud (create, read, update, delete) application.

You had to use the following tech stack for this application: Java, Spring boot and a SQL database and postman to test your API endpoints to ensure that everything worked as it should.

**Note: the controller was the main part of the application. In the controller you will see that the delete function is empty, I was told by the interviewer to leave it empty - they said we know you can do it.**

So then I was asked to refactor my retrieveMovie(...) function, to change it from a string-builder to be a movie object, this is where the Movie.java file is used. (this was an additional part to the interview!).